PyWikibase/requirements.txt
django~=5.1.3
lxml~=5.3.0
django-model-utils~=5.0.0
django-environ~=0.12.0
django-debug-toolbar